updated is_valid_review to consider characters rather words
by counting characters instead of words, it allows us to explicility accept ambiguous or symbol-heavy input that might appear on Letterboxd. a character count is less strict than a "content word" count and it ensures reviews like "great! 10/10" to be marked as valid. stopword-based implementaiton incorrectly filters out short, highly emotional reviews. .isalnum() is one of the fastest ways to process a string in Python.